refactor: remove TreeTrackerDAO and split TreeTackerDAOTest (#1235)
