+++ app/src/main/java/org/greenstand/android/TreeTracker/database/AppDatabase.kt
@@ -77,8 +77,6 @@ abstract class AppDatabase : RoomDatabase() {
 
     abstract fun locationDao(): LocationDAO
 
-    abstract fun treeTrackerDao(): TreeTrackerDAO
-
     companion object {
         private var INSTANCE: AppDatabase? = null
