+++ app/src/main/java/org/greenstand/android/TreeTracker/di/RoomModule.kt
@@ -38,8 +38,6 @@ val roomModule =
 
         single { AppDatabase.getInstance(get()).locationDao() }
 
-        single { AppDatabase.getInstance(get()).treeTrackerDao() }
-
         single { MessageDatabase.getInstance(get()) }
 
         single { MessageDatabase.getInstance(get()).messagesDao() }
